\name{geoEPR}
\docType{data}
\alias{geoEPR}
\title{Polygon dataset to illustrate the functionality of geomerge}
\description{
  \emph{geoEPR} Nigeria dataset used as example for a \code{SpatialPolygonsDataFrame} available from https://icr.ethz.ch/data/epr/geoepr/. The dataset contains geo-locations for all politically relevant ethnic groups from the EPR-Core 2014 dataset. It assigns every politically relevant group one of six settlement patterns and provides polygons describing their location.
}
\usage{data(geomerge)}
\format{A \code{SpatialPolygonsDataFrame} containing observations.}
\source{https://icr.ethz.ch/data/epr/geoepr/

Citation:
Julian Wucherpfennig, Nils B. Weidmann, Luc Girardin, Lars-Erik Cederman, and Andreas Wimmer. (2011). "Politically Relevant Ethnic Groups Across Space and Time: Introducing the GeoEPR Dataset." \emph{Conflict Management and Peace Science} 28(5): 423-437.}
\references{
Andrew M. Linke, Karsten Donnay. (2017). "Scale Variability Misclassification: The Impact of Spatial Resolution on Effect Estimates in the Geographic Analysis of Foreign Aid and Conflict." Paper presented at the \emph{International Studies Association Annual Meeting}, February 22-25 2017, Baltimore.
}
\author{
  Karsten Donnay and Andrew M. Linke
}