\name{gpw}
\docType{data}
\alias{gpw}
\title{Raster dataset to illustrate the functionality of geomerge}
\description{
\emph{gpw} population raster data for Nigeria for the year 2010 used as example for a \code{RasterLayer} available from http://sedac.ciesin.columbia.edu/data/collection/gpw-v4. The dataset (gpw-v4) provides population estimates at a grid resolution of about 4km.
}
\usage{data(geomerge)}
\format{A \code{RasterLayer} containing observations.}
\source{http://sedac.ciesin.columbia.edu/data/collection/gpw-v4

Citation:
Center for International Earth Science Information Network - CIESIN - Columbia University. (2016). \emph{Gridded Population of the World, Version 4 (GPWv4): Population Density}. Palisades, NY: NASA Socioeconomic Data and Applications Center (SEDAC).
}
\references{
Andrew M. Linke, Karsten Donnay. (2017). "Scale Variability Misclassification: The Impact of Spatial Resolution on Effect Estimates in the Geographic Analysis of Foreign Aid and Conflict." Paper presented at the \emph{International Studies Association Annual Meeting}, February 22-25 2017, Baltimore.
}
\author{
  Karsten Donnay and Andrew M. Linke
}