\name{print.geomerge}
\alias{print.geomerge}
\title{
Print function for objects of class 'geomerge'. 
}
\description{
Overloads the default \code{print} for objects of class 'geomerge'.
}
\usage{
\method{print}{geomerge}(x, ...)
}
\arguments{
  \item{x}{
  object of class \code{\link{geomerge}}.}
  \item{...}{
   further arguments passed to or from other methods.}
}
\references{
Andrew M. Linke, Karsten Donnay. (2017). "Scale Variability Misclassification: The Impact of Spatial Resolution on Effect Estimates in the Geographic Analysis of Foreign Aid and Conflict." Paper presented at the \emph{International Studies Association Annual Meeting}, February 22-25 2017, Baltimore.
}

\author{
  Karsten Donnay and Andrew M. Linke.
}
\seealso{
  \code{\link{geomerge}}
}